requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.0.2
Flask-Cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0
SQLAlchemy==1.4.50
Werkzeug==3.0.1
pg8000==1.30.5
psycopg2